Move CURRENT_DISPATCH to tokio-trace-core
